% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abundance.R
\name{aggregate_dfs}
\alias{aggregate_dfs}
\title{Title}
\usage{
aggregate_dfs(iso_abun, group_length, total_length)
}
\arguments{
\item{iso_abun}{}

\item{group_length}{Number of experiments in each repetition. First column contains
names of ions}

\item{total_length}{Number of all experiments}
}
\description{
Title
}
