% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bio_analyzier.R
\name{media_comp_plot}
\alias{media_comp_plot}
\title{Title}
\usage{
media_comp_plot(
  analytes = c("GLC3B", "GLN2B", "GLU2B", "LAC2B"),
  X_sign = c(TRUE, TRUE, FALSE, FALSE),
  rep_num = 2,
  analytes_baseline = c(1, 1, 1, 1),
  mM_baseline = c(1, 1),
  plt_x_axis = c("10mM_VC", "10mM_0.5", "10mM_1", "10mM_2.5", "1mM_VC", "1mM_0.5",
    "1mM_1", "1mM_2.5")
)
}
\arguments{
\item{plt_x_axis}{}
}
\description{
Title
}
