% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abundance.R
\name{multiply_dfs}
\alias{multiply_dfs}
\title{Multiply the ion_count and iso_abun dfs}
\usage{
multiply_dfs(ion_count, iso_abun)
}
\arguments{
\item{ion_count}{ion count data frame}

\item{iso_abun}{mass isotopomer data frame. First column contains all Fragment
names}
}
\value{
df of the multiplication result
}
\description{
Multiply the ion_count and iso_abun dfs
}
