% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abundance.R
\name{plot_abundance}
\alias{plot_abundance}
\title{Plot the stacked barplot representing the abundance}
\usage{
plot_abundance(ion_count, iso_abun, group_length = 8, total_length = 24)
}
\arguments{
\item{ion_count}{ion count data frame}

\item{iso_abun}{mass isotopomer data frame. First column contains all Fragment
names}

\item{group_length}{Number of experiments in each repetition. First column contains
names of ions}

\item{total_length}{Number of all experiments}
}
\value{
plots stored in output folder
}
\description{
Plot the stacked barplot representing the abundance
}
