% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abundance.R
\name{plot_multi_abundance}
\alias{plot_multi_abundance}
\title{Title}
\usage{
plot_multi_abundance(
  col_groups,
  ion_count,
  iso_abun,
  folder_name,
  group_length = 8,
  total_length = 24,
  xtick = NA,
  plot_offset = NA
)
}
\arguments{
\item{offset}{}
}
\description{
Title
}
