% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abundance.R
\name{split_string_take_first}
\alias{split_string_take_first}
\title{Splitting strings in a list of strings by a character, and
taking the first component}
\usage{
split_string_take_first(l, by = " ")
}
\arguments{
\item{l}{list of strings}

\item{by}{character to split strings by}
}
\value{
List of strings
}
\description{
Splitting strings in a list of strings by a character, and
taking the first component
}
\examples{
split_string_take_first(c("abc def", "ghi, jkl"))
}
